requirements.txt
python-telegram-bot==20.7
pandas==2.2.2
openpyxl==3.1.5
selenium==4.23.1
